refactor: 체스판 뷰 변환을 위한 클래스 패키지 이동
